pyproject.toml
[tool.poetry]
name = "task-management-system"
version = "1.0.0"
description = "Tasks For Users"
authors = ["WelmyratDev <ruthlessfront@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
sqlalchemy = ">=2.0.42,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
alembic = ">=1.8.0,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
aiosqlite = ">=0.21.0,<0.22.0"

passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "==3.2.2"

pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
